Most exercises for week 5
For exercise 1.29, I added Simpson's method to calculus.scm,
as another function to pass to integrate-with (along with
rectangle and trapezoid). Because of that framework, I
implemented it by evaluating f at either end of the interval,
and in the middle. As a result, it redundantly evaluates f
at the end of one interval and the beginning of the next,
rather than evaluating it once and multiplying by 2.
Also, if you pass N as the number of intervals to integrate-with,
that corresponds to the book's version of Simpson's method
with n = 2N.